+++ templates/header.inc.php
@@ -293,8 +293,8 @@ $location = get_location();
                 </div>
                 <div id="topmenu_item">
                     <a href="<?php echo AmpConfig::get('web_path'); ?>/stats.php?action=userflag">
-                        <img src="<?php echo $web_path; ?>/images/topmenu-flag.png" />
-                        <span><?php echo T_('My Flags'); ?></span>
+                        <img src="<?php echo $web_path; ?>/images/topmenu-favorite.png" />
+                        <span><?php echo T_('Favorites'); ?></span>
                     </a>
                 </div>
             </div>
@@ -309,7 +309,7 @@ $location = get_location();
         <!-- Tiny little iframe, used to cheat the system -->
         <div id="ajax-loading">Loading . . .</div>
         <iframe name="util_iframe" id="util_iframe" style="display:none;" src="<?php echo AmpConfig::get('web_path'); ?>/util.php"></iframe>
-        <div id="content" class="content-<?php echo AmpConfig::get('ui_fixed') ? 'fixed' : 'float'; ?> <?php echo $count_temp_playlist ? '' : 'content-wild'; ?>">
+        <div id="content" class="content-<?php echo AmpConfig::get('ui_fixed') ? 'fixed' : 'float'; ?> <?php echo (($count_temp_playlist || AmpConfig::get('play_type') == 'localplay') ? '' : 'content-wild'); ?>">
 
         <?php if (AmpConfig::get('int_config_version') != AmpConfig::get('config_version') AND $GLOBALS['user']->has_access(100)) { ?>
         <div class="fatalerror">