+++ templates/show_albums.inc.php
@@ -41,7 +41,7 @@ $thcount = 8;
             <th class="cel_rating optional"><?php echo T_('Rating'); ?></th>
         <?php } ?>
         <?php if (AmpConfig::get('userflags')) { ++$thcount; ?>
-            <th class="cel_userflag optional"><?php echo T_('Flag'); ?></th>
+            <th class="cel_userflag optional"><?php echo T_('Fav.'); ?></th>
         <?php } ?>
             <th class="cel_action essential"><?php echo T_('Actions'); ?></th>
         </tr>
@@ -83,7 +83,7 @@ $thcount = 8;
             <th class="cel_rating"><?php echo T_('Rating'); ?></th>
         <?php } ?>
         <?php if (AmpConfig::get('userflags')) { ?>
-            <th class="cel_userflag"><?php echo T_('Flag'); ?></th>
+            <th class="cel_userflag"><?php echo T_('Fav.'); ?></th>
         <?php } ?>
             <th class="cel_action"><?php echo T_('Actions'); ?></th>
         </tr>