+++ templates/show_now_playing_row.inc.php
@@ -94,7 +94,7 @@ $(document).ready(function(){
         </div>
     </div>
     <div class="np_cell cel_userflag">
-        <label><?php echo T_('Flag'); ?></label>
+        <label><?php echo T_('Fav.'); ?></label>
         <div id="userflag_<?php echo $media->id; ?>_song">
             <?php Userflag::show($media->id,'song'); ?>
         </div>