+++ templates/show_userflag.inc.php
@@ -20,6 +20,6 @@
  *
  */
 ?>
-<?php UI::show_box_top(T_('User Flags')); ?>
+<?php UI::show_box_top(T_('User Favorites')); ?>
 <?php require_once AmpConfig::get('prefix') . '/templates/show_stats_userflag.inc.php'; ?>
 <?php UI::show_box_bottom(); ?>