+++ templates/sidebar_home.inc.php
@@ -81,7 +81,7 @@
       <li id="sb_home_info_Highest"><a href="<?php echo $web_path; ?>/stats.php?action=highest"><?php echo T_('Top Rated'); ?></a></li>
 <?php } ?>
 <?php if (AmpConfig::get('userflags')) { ?>
-      <li id="sb_home_info_UserFlag"><a href="<?php echo $web_path; ?>/stats.php?action=userflag"><?php echo T_('My Flags'); ?></a></li>
+      <li id="sb_home_info_UserFlag"><a href="<?php echo $web_path; ?>/stats.php?action=userflag"><?php echo T_('Favorites'); ?></a></li>
 <?php } ?>
 <?php if (AmpConfig::get('wanted')) { ?>
       <li id="sb_home_info_Wanted"><a href="<?php echo $web_path; ?>/stats.php?action=wanted"><?php echo T_('Wanted List'); ?></a></li>